chore(channel): Move channel part of io service to channel module
