+++ tonic/src/transport/channel/service/io.rs
@@ -1,0 +1,67 @@
+use std::io::{self, IoSlice};
+use std::pin::Pin;
+use std::task::{Context, Poll};
+
+use hyper::rt;
+use hyper_util::client::legacy::connect::{Connected as HyperConnected, Connection};
+
+pub(in crate::transport) trait Io:
+    rt::Read + rt::Write + Send + 'static
+{
+}
+
+impl<T> Io for T where T: rt::Read + rt::Write + Send + 'static {}
+
+pub(crate) struct BoxedIo(Pin<Box<dyn Io>>);
+
+impl BoxedIo {
+    pub(in crate::transport) fn new<I: Io>(io: I) -> Self {
+        BoxedIo(Box::pin(io))
+    }
+}
+
+impl Connection for BoxedIo {
+    fn connected(&self) -> HyperConnected {
+        HyperConnected::new()
+    }
+}
+
+impl rt::Read for BoxedIo {
+    fn poll_read(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        buf: rt::ReadBufCursor<'_>,
+    ) -> Poll<io::Result<()>> {
+        Pin::new(&mut self.0).poll_read(cx, buf)
+    }
+}
+
+impl rt::Write for BoxedIo {
+    fn poll_write(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        buf: &[u8],
+    ) -> Poll<io::Result<usize>> {
+        Pin::new(&mut self.0).poll_write(cx, buf)
+    }
+
+    fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
+        Pin::new(&mut self.0).poll_flush(cx)
+    }
+
+    fn poll_shutdown(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
+        Pin::new(&mut self.0).poll_shutdown(cx)
+    }
+
+    fn poll_write_vectored(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        bufs: &[IoSlice<'_>],
+    ) -> Poll<Result<usize, io::Error>> {
+        Pin::new(&mut self.0).poll_write_vectored(cx, bufs)
+    }
+
+    fn is_write_vectored(&self) -> bool {
+        self.0.is_write_vectored()
+    }
+}
